+++ win-linux/src/csinglewindowbase.cpp
@@ -159,18 +159,18 @@ CSingleWindowBase::CSingleWindowBase(QRect& rect)
     }
 }
 
-void CSingleWindowBase::setScreenScalingFactor(int f)
+void CSingleWindowBase::setScreenScalingFactor(double f)
 {
     if ( m_dpiRatio != f ) {
         if ( isCustomWindowStyle() ) {
-            QSize small_btn_size(TOOLBTN_WIDTH*f, TOOLBTN_HEIGHT*f);
+            QSize small_btn_size(int(TOOLBTN_WIDTH * f), int(TOOLBTN_HEIGHT*f));
 
             m_buttonMinimize->setFixedSize(small_btn_size);
             m_buttonMaximize->setFixedSize(small_btn_size);
             m_buttonClose->setFixedSize(small_btn_size);
 
-            m_boxTitleBtns->setFixedHeight(TOOLBTN_HEIGHT * f);
-            m_boxTitleBtns->layout()->setSpacing(1 * f);
+            m_boxTitleBtns->setFixedHeight(int(TOOLBTN_HEIGHT * f));
+            m_boxTitleBtns->layout()->setSpacing(int(1 * f));
         }
 
 //        onScreenScalingFactor(f);
